Added two method to create an Integer Object and a Double Object in JNI. Method changed with Integer and Double parameters and updated method's signature.

+++ app/src/main/java/it/stefanocappa/ndkexample/Example.java
@@ -78,7 +78,7 @@ public class Example extends AppCompatActivity {
     /**
      * Visibility in not important when i want to call this method from c files
      */
-    private String testMethodWithManyParameters(int a, String b, int c, float d, int e, String[] f, Object g) {
+    private String testMethodWithManyParameters(int a, String b, Integer c, float d, Double e, String[] f, Object g) {
         Log.d(TAG, a + "," + b + "," + c + "," + d + "," + e + "," + f[0]+f[1] + "," + g.toString());
         return a + "," + b + "," + c + "," + d + "," + e + "," + f[0]+f[1] + "," + g.toString();
     }
